requirements.txt
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==45.0.6
daphne==4.2.1
Django==5.2.4
django-cloudinary-storage==0.3.0
django-extensions==4.1
django-filter==25.1
django-ipware==7.0.1
django-phonenumber-field==8.1.0
django-redis==6.0.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.8.1
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
load-dotenv==0.1.0
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==25.0
phonenumberslite==9.0.10
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-ipware==3.0.0
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
service-identity==24.2.0
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
zope.interface==7.2
